requirements/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.1
pillow==10.4.0
sqlparse==0.5.1
